feat: first skelethon winthout accounts and start with config filevalidation
