+++ packages/rpc-proxy/src/utils/args-validator/args-validator-and-getter.ts
@@ -1,14 +1,18 @@
 import { type Config } from '../../types';
 import { type OptionValues } from 'commander';
-import { getConfigObjectFromFile } from '../configuration-file';
+import {
+    checkValidConfigurationFile,
+    getConfigObjectFromFile
+} from '../config-validator';
 import { ArgsValidator } from './args-validator';
-import { InvalidCommandLineArguments } from '@vechain/sdk-errors';
 
 /**
  * Main args validator AND getter object.
  *
  * It validates the command line arguments.
  * AND then it gets a fresh configuration with the validated arguments.
+ *
+ * ArgsValidatorAndGetter is used to validate and, then, get the configuration fields.
  */
 const ArgsValidatorAndGetter = {
     /**
@@ -24,11 +28,15 @@ const ArgsValidatorAndGetter = {
         options: OptionValues,
         defaultConfiguration: Config
     ): Config => {
-        const field = ArgsValidator.configurationFile(
+        const configurationFile = ArgsValidator.configurationFile(
             options.configurationFile as string
         );
-        if (field !== null) {
-            return getConfigObjectFromFile(field);
+        if (configurationFile !== null) {
+            // Check if the configuration file is valid JSON and if it exists
+            checkValidConfigurationFile(configurationFile);
+
+            // Return the configuration object from the configuration file
+            return getConfigObjectFromFile(configurationFile);
         }
         return defaultConfiguration;
     },
@@ -53,6 +61,10 @@ const ArgsValidatorAndGetter = {
     },
 
     /**
+     * ********* START: TEMPORARY COMMENT *********
+     * This method will be implemented in the future.
+     * ********* END: TEMPORARY COMMENT ********
+     *
      * Validate 'url' configuration field
      *
      * @param options Command line arguments options
@@ -69,9 +81,13 @@ const ArgsValidatorAndGetter = {
             } satisfies Config;
         }
         return currentConfiguration;
-    },
+    }
 
     /**
+     * ********* START: TEMPORARY COMMENT *********
+     * This method will be implemented in the future.
+     * ********* END: TEMPORARY COMMENT ********
+     *
      * Validate 'accounts' configuration field
      *
      * @param options Command line arguments options
@@ -79,18 +95,22 @@ const ArgsValidatorAndGetter = {
      * @returns Configuration object
      * @throws {InvalidCommandLineArguments}
      */
-    accounts: (options: OptionValues, currentConfiguration: Config): Config => {
-        const field = ArgsValidator.accounts(options.accounts as string);
-        if (field !== null) {
-            return {
-                ...currentConfiguration,
-                accounts: field
-            } satisfies Config;
-        }
-        return currentConfiguration;
-    },
+    // accounts: (options: OptionValues, currentConfiguration: Config): Config => {
+    //     const field = ArgsValidator.accounts(options.accounts as string);
+    //     if (field !== null) {
+    //         return {
+    //             ...currentConfiguration,
+    //             accounts: field
+    //         } satisfies Config;
+    //     }
+    //     return currentConfiguration;
+    // }
 
     /**
+     * ********* START: TEMPORARY COMMENT *********
+     * This method will be implemented in the future.
+     * ********* END: TEMPORARY COMMENT ********
+     *
      * Validate 'mnemonic' configuration field
      *
      * @param options Command line arguments options
@@ -98,61 +118,122 @@ const ArgsValidatorAndGetter = {
      * @returns Configuration object
      * @throws {InvalidCommandLineArguments}
      */
-    mnemonic: (options: OptionValues, currentConfiguration: Config): Config => {
-        const field = ArgsValidator.mnemonic(options.mnemonic as string);
+    // mnemonic: (options: OptionValues, currentConfiguration: Config): Config => {
+    //     const mnemonicToUse = ArgsValidator.mnemonic(
+    //         options.mnemonic as string
+    //     );
+    //
+    //     // Mnemonic count must be provided if mnemonic is provided
+    //     if (
+    //         mnemonicToUse !== null &&
+    //         (options.mnemonicCount === undefined ||
+    //             options.mnemonicCount === null)
+    //     ) {
+    //         throw new InvalidCommandLineArguments(
+    //             'ArgsValidatorAndGetter.mnemonic()',
+    //             'No mnemonic count provided. A mnemonic count must be provided',
+    //             {
+    //                 flag: '-mc , --mnemonicCount',
+    //                 value: 'not provided'
+    //             }
+    //         );
+    //     }
+    //
+    //     // Mnemonic initial index must be provided if mnemonic is provided
+    //     if (
+    //         mnemonicToUse !== null &&
+    //         (options.mnemonicInitialIndex === undefined ||
+    //             options.mnemonicInitialIndex === null)
+    //     ) {
+    //         throw new InvalidCommandLineArguments(
+    //             'ArgsValidatorAndGetter.mnemonic()',
+    //             'No mnemonic initial index provided. A mnemonic initial index must be provided',
+    //             {
+    //                 flag: '-mi , --mnemonicInitialIndex',
+    //                 value: 'not provided'
+    //             }
+    //         );
+    //     }
+    //
+    //     // @note: This field must be provided otherwise previous validation fails!
+    //     if (
+    //         options.mnemonicCount !== undefined &&
+    //         options.mnemonicCount !== null &&
+    //         options.mnemonicInitialIndex !== undefined &&
+    //         options.mnemonicInitialIndex !== null
+    //     ) {
+    //         const field2 = ArgsValidator.mnemonicCount(
+    //             options.mnemonicCount as string
+    //         );
+    //
+    //         // @note: This field must be provided otherwise previous validation fails!
+    //         const field3 = ArgsValidator.mnemonicInitialIndex(
+    //             options.mnemonicInitialIndex as string
+    //         );
+    //
+    //         if (mnemonicToUse !== null) {
+    //             return {
+    //                 ...currentConfiguration,
+    //                 accounts: {
+    //                     mnemonic: mnemonicToUse,
+    //                     count: field2,
+    //                     initialIndex: field3
+    //                 }
+    //             } satisfies Config;
+    //         }
+    //     }
+    //     return currentConfiguration;
+    // }
 
-        // Mnemonic count must be provided if mnemonic is provided
-        if (
-            (field !== null && options.mnemonicCount === undefined) ||
-            options.mnemonicCount === null
-        ) {
-            throw new InvalidCommandLineArguments(
-                'ArgsValidatorAndGetter.mnemonic()',
-                'No mnemonic count provided. A mnemonic count must be provided',
-                {
-                    flag: '-mc , --mnemonicCount',
-                    value: 'not provided'
-                }
-            );
-        }
-
-        // Mnemonic initial index must be provided if mnemonic is provided
-        if (
-            (field !== null && options.mnemonicInitialIndex === undefined) ||
-            options.mnemonicInitialIndex === null
-        ) {
-            throw new InvalidCommandLineArguments(
-                'ArgsValidatorAndGetter.mnemonic()',
-                'No mnemonic initial index provided. A mnemonic initial index must be provided',
-                {
-                    flag: '-mi , --mnemonicInitialIndex',
-                    value: 'not provided'
-                }
-            );
-        }
-
-        // @note: This field must be provided otherwise previous validation fails!
-        const field2 = ArgsValidator.mnemonicCount(
-            options.mnemonicCount as string
-        );
-
-        // @note: This field must be provided otherwise previous validation fails!
-        const field3 = ArgsValidator.mnemonicInitialIndex(
-            options.mnemonicInitialIndex as string
-        );
-
-        if (field !== null) {
-            return {
-                ...currentConfiguration,
-                accounts: {
-                    mnemonic: field,
-                    count: field2,
-                    initialIndex: field3
-                }
-            } satisfies Config;
-        }
-        return currentConfiguration;
-    }
+    /**
+     * ********* START: TEMPORARY COMMENT *********
+     * This method will be implemented in the future.
+     * ********* END: TEMPORARY COMMENT ********
+     *
+     * Validate 'delegation' configuration fields
+     *
+     * @param options Command line arguments options
+     * @param currentConfiguration Default configuration to use if no field is provided
+     * @returns Configuration object
+     * @throws {InvalidCommandLineArguments}
+     */
+    // delegation: (
+    //     options: OptionValues,
+    //     currentConfiguration: Config
+    // ): Config => {
+    //     // Both delegation fields are provided - throw an error
+    //     if (
+    //         options.delegatorPrivateKey !== undefined &&
+    //         options.delegatorUrl !== undefined &&
+    //         options.delegatorPrivateKey !== null &&
+    //         options.delegatorUrl !== null
+    //     ) {
+    //         throw new InvalidCommandLineArguments(
+    //             'ArgsValidatorAndGetter.delegation()',
+    //             'Both delegator private key and delegator URL are provided. Only one can be provided',
+    //             {
+    //                 flag: '{-dp , --delegatorPrivateKey}, {-du , --delegatorUrl}',
+    //                 value: `{value not provided for security reason} , {${options.delegatorUrl as string}}`
+    //             }
+    //         );
+    //     }
+    //
+    //     // Delegation is made with a private key
+    //     if (
+    //         options.delegatePrivateKey !== undefined &&
+    //         options.delegatePrivateKey !== null
+    //     ) {
+    //     }
+    //
+    //     // Delegation is made with a delegator URL
+    //     if (
+    //         options.delegatorUrl !== undefined &&
+    //         options.delegatorUrl !== null
+    //     ) {
+    //     }
+    //
+    //     return currentConfiguration;
+    // }
 };
 
 export { ArgsValidatorAndGetter };
