+++ packages/rpc-proxy/src/utils/args-validator/args-validator.ts
@@ -1,26 +1,13 @@
-import { checkValidConfigurationFile } from '../configuration-file';
 import { InvalidCommandLineArguments } from '@vechain/sdk-errors';
-import { HDNode, Hex, secp256k1 } from '@vechain/sdk-core';
-
-/**
- * Check if the url is valid function
- * @param url - URL to check
- * @returns True if the url is valid, false otherwise
- */
-const isValidUrl = (url: string): boolean => {
-    try {
-        // eslint-disable-next-line no-new
-        new URL(url);
-        return true;
-    } catch {
-        return false;
-    }
-};
+import { checkValidConfigurationFile } from '../config-validator';
+import { isValidPort, isValidUrl } from '../validators';
 
 /**
  * Main args validator.
  *
  * It validates the command line arguments.
+ *
+ * ArgsValidator is used to validate every single field of the configuration.
  */
 const ArgsValidator = {
     /**
@@ -38,11 +25,8 @@ const ArgsValidator = {
             configurationFilePath !== undefined &&
             configurationFilePath !== null
         ) {
-            // Check if the configuration file exists AND is a valid JSON
+            // Check if the configuration file is valid: It exists and is a valid JSON
             checkValidConfigurationFile(configurationFilePath);
-
-            // Check the semantic of the configuration file
-            // TODO
 
             console.log(
                 `[rpc-proxy]: Configuration file provided ${configurationFilePath}`
@@ -68,14 +52,10 @@ const ArgsValidator = {
         if (port !== undefined && port !== null) {
             const portAsNumber = Number(port.toString());
 
-            if (
-                isNaN(portAsNumber) ||
-                !Number.isInteger(portAsNumber) ||
-                portAsNumber < 0
-            ) {
+            if (!isValidPort(portAsNumber)) {
                 throw new InvalidCommandLineArguments(
                     'ArgsValidator.port()',
-                    'Invalid port provided. A port must be an integer',
+                    'Invalid port provided. The parameter must be an integer',
                     {
                         flag: '-p , --port',
                         value: port
@@ -125,136 +105,225 @@ const ArgsValidator = {
             );
         }
         return null;
-    },
+    }
 
     /**
+     * ********* START: TEMPORARY COMMENT *********
+     * This method will be implemented in the future.
+     * ********* END: TEMPORARY COMMENT ********
+     *
      * Validate 'accounts' configuration field
      *
      * @param accounts Accounts to validate
      * @returns Accounts as a list of private keys, if provided AND valid, null otherwise
      * @throws {InvalidCommandLineArguments}
      */
-    accounts: (accounts?: string | null): string[] | null => {
-        if (accounts !== undefined && accounts !== null) {
-            const accountsList: string[] = accounts.split(' ');
+    // accounts: (accounts?: string | null): string[] | null => {
+    //     if (accounts !== undefined && accounts !== null) {
+    //         const accountsList: string[] = accounts.split(' ');
+    //
+    //         accountsList
+    //             .map(
+    //                 (accountPrivateKeyAsString) =>
+    //                     Hex.of(accountPrivateKeyAsString).bytes
+    //             )
+    //             .forEach((privateKey: Uint8Array) => {
+    //                 if (
+    //                     !secp256k1.isValidPrivateKey(Hex.of(privateKey).bytes)
+    //                 ) {
+    //                     throw new InvalidCommandLineArguments(
+    //                         'ArgsValidator.accounts()',
+    //                         'An invalid account private key provided.',
+    //                         {
+    //                             flag: '-a , --accounts',
+    //                             value: 'Value will not be shown for security reasons'
+    //                         }
+    //                     );
+    //                 }
+    //             });
+    //
+    //         return accountsList;
+    //     } else {
+    //         console.log(
+    //             '[rpc-proxy]: No accounts provided with command line arguments. Default port will be used.'
+    //         );
+    //     }
+    //     return null;
+    // }
 
-            accountsList
-                .map(
-                    (accountPrivateKeyAsString) =>
-                        Hex.of(accountPrivateKeyAsString).bytes
-                )
-                .forEach((privateKey: Uint8Array) => {
-                    if (
-                        !secp256k1.isValidPrivateKey(Hex.of(privateKey).bytes)
-                    ) {
-                        throw new InvalidCommandLineArguments(
-                            'ArgsValidator.accounts()',
-                            'An invalid account private key provided.',
-                            {
-                                flag: '-a , --accounts',
-                                value: 'Value will not be shown for security reasons'
-                            }
-                        );
-                    }
-                });
-
-            return accountsList;
-        } else {
-            console.log(
-                '[rpc-proxy]: No accounts provided with command line arguments. Default port will be used.'
-            );
-        }
-        return null;
-    },
-
-    /**
+    /*
+     * ********* START: TEMPORARY COMMENT *********
+     * This method will be implemented in the future.
+     * ********* END: TEMPORARY COMMENT ********
+     *
      * Validate 'mnemonic' configuration field
      *
      * @param mnemonic Mnemonic to validate
      * @returns Mnemonic if provided AND valid, null otherwise
      * @throws {InvalidCommandLineArguments}
      */
-    mnemonic: (mnemonic?: string | null): string | null => {
-        if (mnemonic !== undefined && mnemonic !== null) {
-            try {
-                HDNode.fromMnemonic(mnemonic.split(' '));
-            } catch (e) {
-                throw new InvalidCommandLineArguments(
-                    'ArgsValidator.mnemonic()',
-                    'An invalid account mnemonic is present in the configuration file',
-                    {
-                        flag: '-m , --mnemonic',
-                        value: 'Value will not be shown for security reasons'
-                    }
-                );
-            }
-            return mnemonic;
-        } else {
-            console.log(
-                '[rpc-proxy]: No mnemonic provided with command line arguments. Default port will be used.'
-            );
-        }
-
-        return null;
-    },
+    // mnemonic: (mnemonic?: string | null): string | null => {
+    //     if (mnemonic !== undefined && mnemonic !== null) {
+    //         try {
+    //             HDNode.fromMnemonic(mnemonic.split(' '));
+    //         } catch (e) {
+    //             throw new InvalidCommandLineArguments(
+    //                 'ArgsValidator.mnemonic()',
+    //                 'An invalid account mnemonic is present in the configuration file',
+    //                 {
+    //                     flag: '-m , --mnemonic',
+    //                     value: 'Value will not be shown for security reasons'
+    //                 }
+    //             );
+    //         }
+    //         return mnemonic;
+    //     } else {
+    //         console.log(
+    //             '[rpc-proxy]: No mnemonic provided with command line arguments. Default port will be used.'
+    //         );
+    //     }
+    //
+    //     return null;
+    // }
 
     /**
+     * ********* START: TEMPORARY COMMENT *********
+     * This method will be implemented in the future.
+     * ********* END: TEMPORARY COMMENT ********
+     *
      * Validate 'mnemonicCount' configuration field
      *
      * @param mnemonicCount Mnemonic count to validate
      * @returns Mnemonic count if provided AND valid, null otherwise
      * @throws {InvalidCommandLineArguments}
      */
-    mnemonicCount: (mnemonicCount: string): number => {
-        const mnemonicCountAsNumber = Number(mnemonicCount.toString());
-
-        if (
-            isNaN(mnemonicCountAsNumber) ||
-            !Number.isInteger(mnemonicCountAsNumber) ||
-            mnemonicCountAsNumber < 0
-        ) {
-            throw new InvalidCommandLineArguments(
-                'ArgsValidator.mnemonicCount()',
-                'Invalid mnemonicCount provided. A port must be an integer',
-                {
-                    flag: '-mc , --mnemonicCount',
-                    value: mnemonicCount
-                }
-            );
-        }
-
-        return mnemonicCountAsNumber;
-    },
+    // mnemonicCount: (mnemonicCount: string): number => {
+    //     const mnemonicCountAsNumber = Number(mnemonicCount.toString());
+    //
+    //     if (
+    //         isNaN(mnemonicCountAsNumber) ||
+    //         !Number.isInteger(mnemonicCountAsNumber) ||
+    //         mnemonicCountAsNumber < 0
+    //     ) {
+    //         throw new InvalidCommandLineArguments(
+    //             'ArgsValidator.mnemonicCount()',
+    //             'Invalid mnemonicCount provided. A port must be an integer',
+    //             {
+    //                 flag: '-mc , --mnemonicCount',
+    //                 value: mnemonicCount
+    //             }
+    //         );
+    //     }
+    //
+    //     return mnemonicCountAsNumber;
+    // }
 
     /**
+     * ********* START: TEMPORARY COMMENT *********
+     * This method will be implemented in the future.
+     * ********* END: TEMPORARY COMMENT ********
+     *
      * Validate 'mnemonicInitialIndex' configuration field
      *
      * @param mnemonicInitialIndex Mnemonic initial index to validate
      * @returns Mnemonic initial index if provided AND valid, null otherwise
      * @throws {InvalidCommandLineArguments}
      */
-    mnemonicInitialIndex: (mnemonicInitialIndex: string): number => {
-        const mnemonicInitialIndexAsNumber = Number(
-            mnemonicInitialIndex.toString()
-        );
+    // mnemonicInitialIndex: (mnemonicInitialIndex: string): number => {
+    //     const mnemonicInitialIndexAsNumber = Number(
+    //         mnemonicInitialIndex.toString()
+    //     );
+    //
+    //     if (
+    //         isNaN(mnemonicInitialIndexAsNumber) ||
+    //         !Number.isInteger(mnemonicInitialIndexAsNumber) ||
+    //         mnemonicInitialIndexAsNumber < 0
+    //     ) {
+    //         throw new InvalidCommandLineArguments(
+    //             'ArgsValidator.mnemonicInitialIndex()',
+    //             'Invalid mnemonicInitialIndex provided. A port must be an integer',
+    //             {
+    //                 flag: '-mi , --mnemonicInitialIndex',
+    //                 value: mnemonicInitialIndex
+    //             }
+    //         );
+    //     }
+    //
+    //     return mnemonicInitialIndexAsNumber;
+    // }
 
-        if (
-            isNaN(mnemonicInitialIndexAsNumber) ||
-            !Number.isInteger(mnemonicInitialIndexAsNumber) ||
-            mnemonicInitialIndexAsNumber < 0
-        ) {
-            throw new InvalidCommandLineArguments(
-                'ArgsValidator.mnemonicInitialIndex()',
-                'Invalid mnemonicInitialIndex provided. A port must be an integer',
-                {
-                    flag: '-mi , --mnemonicInitialIndex',
-                    value: mnemonicInitialIndex
-                }
-            );
-        }
+    /**
+     * ********* START: TEMPORARY COMMENT *********
+     * This method will be implemented in the future.
+     * ********* END: TEMPORARY COMMENT ********
+     *
+     * Delegate configuration
+     * Validate 'delegatorPrivateKey' configuration field.
+     *
+     * @param delegatorPrivateKey Delegator private key to validate
+     * @returns Delegator private key if provided AND valid, null otherwise
+     */
+    // delegatorPrivateKey: (delegatorPrivateKey: string): string | null => {
+    //     if (delegatorPrivateKey !== undefined && delegatorPrivateKey !== null) {
+    //         if (
+    //             !secp256k1.isValidPrivateKey(Hex.of(delegatorPrivateKey).bytes)
+    //         ) {
+    //             throw new InvalidCommandLineArguments(
+    //                 'ArgsValidator.delegatorPrivateKey()',
+    //                 'An invalid delegator private key provided.',
+    //                 {
+    //                     flag: '-dp , --delegatorPrivateKey',
+    //                     value: 'Value will not be shown for security reasons'
+    //                 }
+    //             );
+    //         }
+    //         console.log(
+    //             `[rpc-proxy]: Delegator private key provided with command line options`
+    //         );
+    //
+    //         return delegatorPrivateKey;
+    //     } else {
+    //         console.log(
+    //             '[rpc-proxy]: No url provided with command line arguments. Default port will be used.'
+    //         );
+    //     }
+    //     return null;
+    // }
 
-        return mnemonicInitialIndexAsNumber;
-    }
+    /*
+     * ********* START: TEMPORARY COMMENT *********
+     * This method will be implemented in the future.
+     * ********* END: TEMPORARY COMMENT ********
+     *
+     * Validate 'delgatorUrl' configuration field
+     *
+     * @param delegatorUrl Delegator URL to validate
+     * @returns Delegator URL if provided AND valid, null otherwise
+     */
+    // delegatorUrl: (delegatorUrl: string): string | null => {
+    //     if (delegatorUrl !== undefined && delegatorUrl !== null) {
+    //         if (!isValidUrl(delegatorUrl)) {
+    //             throw new InvalidCommandLineArguments(
+    //                 'ArgsValidator.delegatorUrl()',
+    //                 'Invalid delegator url provided. The parameter must be a valid url',
+    //                 {
+    //                     flag: '-du , --delegatorUrl',
+    //                     value: delegatorUrl
+    //                 }
+    //             );
+    //         }
+    //         console.log(
+    //             `[rpc-proxy]: Delegator url provided with command line options: ${delegatorUrl}`
+    //         );
+    //
+    //         return delegatorUrl;
+    //     } else {
+    //         console.log(
+    //             '[rpc-proxy]: No delegator url provided with command line arguments. Default port will be used.'
+    //         );
+    //     }
+    //     return null;
+    // }
 };
 
 export { ArgsValidator };
