+++ packages/rpc-proxy/src/utils/config-validator/index.ts
@@ -1,0 +1,1 @@
+export * from './config-validator';
