--- packages/rpc-proxy/src/utils/configuration-file/index.ts
@@ -1,1 +1,0 @@
-export * from './configuration-file';
