+++ packages/rpc-proxy/src/utils/index.ts
@@ -1,3 +1,4 @@
 export * from './args';
-export * from './configuration-file';
 export * from './args-validator';
+export * from './config-validator';
+export * from './validators';
